fix: Extract sort hook from sortedTable component
